+++ packages/wallet-sdk/src/sign/interface.ts
@@ -1,4 +1,13 @@
-import { AddressString, IntNumber } from ':core/type';
+import { RequestArguments } from ':core/provider/interface';
+import { AddressString, Chain, IntNumber } from ':core/type';
+
+export interface Signer {
+  readonly accounts: AddressString[];
+  readonly chain: Chain;
+  handshake(): Promise<AddressString[]>;
+  request<T>(request: RequestArguments): Promise<T>;
+  disconnect: () => Promise<void>;
+}
 
 export interface StateUpdateListener {
   onAccountsUpdate: (_: AddressString[]) => void;
